+++ spec/controllers/api/v1/participants_controller_spec.rb
